PTX-2226 Added test for PVC and Storage pool resize at the same time
