+++ deployments/customconfigs/default-workload-bigger-volume.yaml
@@ -1,0 +1,3 @@
+aut-postgres:
+  workload_size: 10
+  volume_size: 20Gi
