+++ pkg/aututils/aututils.go
@@ -2,6 +2,7 @@ package aututils
 
 import (
 	"fmt"
+	"strings"
 
 	apapi "github.com/libopenstorage/autopilot-api/pkg/apis/autopilot/v1alpha1"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -31,6 +32,44 @@ const (
 	// StorageSpecAction is name for storage spec action
 	StorageSpecAction = "openstorage.io.action.storagepool/expand"
 )
+
+// PvcRuleByUsageCapacity returns an autopilot pvc expand rule that uses usage of pvc size
+func PvcRuleByUsageCapacity(usage, scalePercentage int, maxSize string) apapi.AutopilotRule {
+	apRuleName := fmt.Sprintf("pvc-usage-%d-scale-%d", usage, scalePercentage)
+	apRuleSpecActions := []*apapi.RuleAction{
+		{
+			Name: VolumeSpecAction,
+			Params: map[string]string{
+				RuleActionsScalePercentage: fmt.Sprintf("%d", scalePercentage),
+			},
+		},
+	}
+	if maxSize != "" {
+		apRuleName = fmt.Sprintf("%s-maxsize-%s", apRuleName, maxSize)
+		for _, specAction := range apRuleSpecActions {
+			specAction.Params[RuleMaxSize] = maxSize
+		}
+	}
+	apRuleObject := apapi.AutopilotRule{
+		ObjectMeta: meta_v1.ObjectMeta{
+			Name: strings.ToLower(apRuleName),
+		},
+		Spec: apapi.AutopilotRuleSpec{
+			Conditions: apapi.RuleConditions{
+				Expressions: []*apapi.LabelSelectorRequirement{
+					{
+						Key:      PxVolumeUsagePercentMetric,
+						Operator: apapi.LabelSelectorOpGt,
+						Values:   []string{fmt.Sprintf("%d", usage)},
+					},
+				},
+			},
+			Actions: apRuleSpecActions,
+		},
+	}
+
+	return apRuleObject
+}
 
 // PoolRuleByTotalSize returns an autopilot pool expand rule that uses total pool size
 func PoolRuleByTotalSize(total, scalePercentage int, expandType string, labelSelector map[string]string) apapi.AutopilotRule {
